+++ ntsc-mtl-demo.xcodeproj/xcuserdata/jeff.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -16,8 +16,8 @@
             endingColumnNumber = "9223372036854775807"
             startingLineNumber = "24"
             endingLineNumber = "24"
-            landmarkName = "run(inputTexture:outputTexture:commandBuffer:)"
-            landmarkType = "7">
+            landmarkName = "unknown"
+            landmarkType = "0">
          </BreakpointContent>
       </BreakpointProxy>
       <BreakpointProxy
