+++ ntsc-mtl-demo/Filters/MetalKernels/Python/Paint.metal
