+++ ntsc-mtl-demo/Filters/TextureFilters/Python/TextureFilterError.swift
@@ -1,0 +1,22 @@
+//
+//  TextureFilterError.swift
+//  ntsc-mtl-demo
+//
+//  Created by Jeffrey Blagdon on 2024-06-23.
+//
+
+import Foundation
+
+enum TextureFilterError: Swift.Error {
+    case cantMakeTexture
+    case cantMakeCommandQueue
+    case cantMakeCommandBuffer
+    case cantMakeComputeEncoder
+    case cantMakeLibrary
+    case cantMakeRandomImage
+    case cantMakeFilter(String)
+    case cantMakeFunction(String)
+    case cantMakeBlitEncoder
+    case logicHole(String)
+    case notImplemented
+}
